+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/ref/RobotPythonClass.java
@@ -26,6 +26,14 @@ public class RobotPythonClass implements KeywordFile {
         this.importType = importType;
     }
 
+    @Override
+    public Collection<DefinedVariable> findDefinedVariable(@NotNull String variableName) {
+        if (importType == ImportType.VARIABLES) {
+            return RobotKeywordFileResolver.findVariable(pythonClass, variableName);
+        }
+        return List.of();
+    }
+
     @NotNull
     @Override
     public Collection<DefinedVariable> getDefinedVariables(Collection<KeywordFile> visitedFiles) {