+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/ref/RobotVariableBodyReference.java
@@ -1,33 +1,27 @@
 package dev.xeonkryptos.xeonrobotframeworkplugin.psi.ref;
 
-import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementResolveResult;
-import com.intellij.psi.PsiFile;
 import com.intellij.psi.PsiPolyVariantReferenceBase;
 import com.intellij.psi.ResolveResult;
 import com.intellij.psi.impl.source.resolve.ResolveCache;
 import com.intellij.psi.search.GlobalSearchScope;
 import com.intellij.psi.util.PsiTreeUtil;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotFeatureFileType;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.RobotResourceFileType;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.KeywordFile;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.DefinedVariable;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotFile;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotVariable;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotVariableBodyId;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotVariableDefinition;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.stub.index.VariableDefinitionNameIndex;
-import dev.xeonkryptos.xeonrobotframeworkplugin.psi.stub.index.VariableNameIndex;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.util.VariableScope;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.LinkedHashSet;
-import java.util.Map;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 public class RobotVariableBodyReference extends PsiPolyVariantReferenceBase<RobotVariableBodyId> {
 
@@ -79,46 +73,12 @@ public class RobotVariableBodyReference extends PsiPolyVariantReferenceBase<Robo
 
     @NotNull
     private Collection<PsiElement> findVariableElementsOutsideOfCurrentFile(RobotVariable variable, String variableName) {
-        Project project = variable.getProject();
-        GlobalSearchScope globalSearchScope = GlobalSearchScope.projectScope(project);
-        Collection<RobotVariableDefinition> matchingVariableDefinitions = VariableDefinitionNameIndex.getInstance()
-                                                                                                     .getVariableDefinitions(variableName,
-                                                                                                                             project,
-                                                                                                                             globalSearchScope);
-        Collection<RobotVariable> matchingVariables = VariableNameIndex.getInstance().getVariables(variableName, project, globalSearchScope);
-
-        Map<RobotFile, Collection<PsiElement>> fileToElements = new HashMap<>();
-        for (RobotVariableDefinition variableDefinition : matchingVariableDefinitions) {
-            PsiFile containingFile = variableDefinition.getContainingFile();
-            if (containingFile.getFileType() == RobotFeatureFileType.getInstance() || containingFile.getFileType() == RobotResourceFileType.getInstance()) {
-                fileToElements.computeIfAbsent((RobotFile) containingFile, k -> new LinkedHashSet<>()).add(variableDefinition);
-            }
-        }
-        for (RobotVariable matchingVariable : matchingVariables) {
-            PsiFile containingFile = matchingVariable.getContainingFile();
-            if (containingFile.getFileType() == RobotFeatureFileType.getInstance() || containingFile.getFileType() == RobotResourceFileType.getInstance()) {
-                fileToElements.computeIfAbsent((RobotFile) containingFile, k -> new LinkedHashSet<>()).add(matchingVariable);
-            }
-        }
-
-        Set<PsiElement> inScopeElements = new LinkedHashSet<>();
         RobotFile robotFile = (RobotFile) variable.getContainingFile();
-        for (KeywordFile importedFile : robotFile.collectImportedFiles(true)) {
-            PsiFile importedPsiFile = importedFile.getPsiFile();
-            if ((importedPsiFile.getFileType() == RobotFeatureFileType.getInstance() || importedPsiFile.getFileType() == RobotResourceFileType.getInstance())
-                && fileToElements.containsKey((RobotFile) importedPsiFile)) {
-                Collection<PsiElement> psiElements = fileToElements.get(((RobotFile) importedPsiFile));
-                for (PsiElement psiElement : psiElements) {
-                    if (psiElement instanceof RobotVariableDefinition definition) {
-                        if (EASY_SCOPES.contains(definition.getScope()) || definition.isInScope(variable)) {
-                            inScopeElements.add(definition);
-                        }
-                    } else {
-                        inScopeElements.add(psiElement);
-                    }
-                }
-            }
-        }
-        return inScopeElements;
+        return robotFile.collectImportedFiles(true)
+                        .stream()
+                        .flatMap(file -> file.findDefinedVariable(variableName).stream())
+                        .filter(foundVar -> foundVar.matches(variableName) && (EASY_SCOPES.contains(foundVar.getScope()) || foundVar.isInScope(variable)))
+                        .map(DefinedVariable::reference)
+                        .collect(Collectors.toCollection(LinkedHashSet::new));
     }
 }