+++ src/main/kotlin/dev/xeonkryptos/xeonrobotframeworkplugin/util/RobotNames.kt
@@ -3,4 +3,6 @@ package dev.xeonkryptos.xeonrobotframeworkplugin.util
 object RobotNames {
 
     const val INIT_DOT_ROBOT = "__init__.robot"
+    const val LIST_VARIABLE_TYPE_INDICATOR_PREFIX = "LIST__"
+    const val DICT_VARIABLE_TYPE_INDICATOR_PREFIX = "DICT__"
 }