+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     - if not found, the imported file is looked up relative to the content root of the module search paths. In case of a library or variables import, the lookup
       looks also for classes and modules
 - Use index lookup for finding referenced variable definitions in the same file to speed up the process
+- Use index lookup for finding referenced variable definitions in imported python modules and classes to speed up the process
 
 ## 0.5.11